+++ index.js
@@ -3,5 +3,6 @@ module.exports = {
   VrpcNative: require('./vrpc/VrpcNative'),
   VrpcClient: require('./vrpc/VrpcClient'),
   VrpcAdapter: require('./vrpc/VrpcAdapter'),
-  VrpcAgent: require('./vrpc/VrpcAgent')
+  VrpcAgent: require('./vrpc/VrpcAgent'),
+  VrpcPersistor: require('./vrpc/VrpcPersistor')
 }