+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.3.0] - Sep 05 2026
+
+### Added
+
+- new class `VrpcPersistor` providing a persistence layer that can automatically
+  save and restore instances
+
 ## [3.2.1] - Jun 26 2025
 
 ### Fixed